requirements.txt
fastapi
uvicorn
openai
python-dotenv
websockets
pydantic
numpy
scipy
librosa
speechrecognition
pyaudio
python-multipart
aiofiles
pypdf2
python-docx
matplotlib
seaborn
plotly
jinja2
websocket-client
requests